Allow PatternElement strings to be borrowed
The strings in PatternElement only need to be re-allocated when they are first
stored on a Clusterer. When they are ephemeral patterns used on a single line,
it's fine to use borrowed string slices. This significantly reduces allocation
pressure and improves performance.

After applying this commit, I saw throughput jump from ~5MiB/s to ~9MiB/s.

I also was able to reduce allocations by re-using vector heap space in
`create_pattern`. This introduced some unsafe code that I'm not extremely happy
with. I opened [this stackoverflow question][1] to see if there is a better way
to do this, but for now it works.

[1]: https://stackoverflow.com/questions/69112758