Significantly speed up parallel mode and make it default
Previously, parallel mode was not much faster (and was significantly slower on
linux) than the single-core mode. This was due to a poor read distribution
strategy. I had a single IO thread that pulled lines from stdin and distributed
these lines to worker threads. The new strategy performs IO on each individual
worker thread, using StdinLocks to ensure that they do not mess with the cursor
position. I also added line chunking so that each thread reads a certain number
of lines at a time before releasing the lock. This should reduce the locking
overhead by reducing the total number of times locks need to be taken, and
allows lines to queue up in memory on a particular worker thread.